requirements.txt
#ansible==2.10.7
ansible-runner==2.0.1
boto3==1.17.95
boto3_type_annotations==0.3.1
click==8.0.1
emoji==1.2.0
Jinja2==2.10.3
progressbar2==3.53.1
pytest==6.2.4
pytest-dependency==0.5.1
pytest-order==1.0.0
PyYAML==5.3.1
validators==0.18.2
